% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCathRes.R
\name{calculateCathodeResMean}
\alias{calculateCathodeResMean}
\title{calculateCathoResMean}
\usage{
calculateCathodeResMean(.d, .group, .age, .res, .ageRange = 1:2000)
}
\arguments{
\item{.d}{data.frame with data}

\item{.group}{char with group names}

\item{.age}{cell age}

\item{.res}{cathode resistance values}

\item{.ageRange}{range of age for the calculation of the mean}
}
\value{
dataframe with 2 columns, name of group and average resistance for
each group
}
\description{
Calculates the average of groups of cathode resistance values
based on a + b.log(age) model for each group
}
