% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateResDiff.R
\name{calculateResDiff}
\alias{calculateResDiff}
\title{calculateResDiff : calculates resistance difference}
\usage{
calculateResDiff(d, group, age, res)
}
\arguments{
\item{d}{data.frame with data}

\item{group}{char with group names}

\item{age}{cell age}

\item{res}{cathode resistance value}
}
\value{
summary of model using broom::tidy
}
\description{
Calculates the difference in cathode resistance between cathode resistance groups assuming
the same evolution with age and 0 intercept
}
