% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCathRes.R
\name{modelCathoRes1}
\alias{modelCathoRes1}
\title{modelCathoRes1}
\usage{
modelCathoRes1(d, group, age, res)
}
\arguments{
\item{d}{data.frame with data}

\item{group}{char with group names}

\item{age}{cell age}

\item{res}{cathode resistance value}
}
\value{
summary of model using broom::tidy
}
\description{
Model cathode resistance per groups assuming
the same evolution with age and 0 intercept - using lme4
}
