% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCathRes.R
\name{modelCathoResLm}
\alias{modelCathoResLm}
\title{modelCathoLm}
\usage{
modelCathoResLm(d, group, age, res)
}
\arguments{
\item{d}{data.frame with data}

\item{group}{char with group names}

\item{age}{cell age}

\item{res}{cathode resistance value}
}
\value{
summary of model using broom::tidy
}
\description{
Model cathode resistance per groups assuming
evolution with log of age and 0 fixed intercept - using lme4
}
