% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCathRes.R
\name{modelCathoResLme}
\alias{modelCathoResLme}
\title{modelCathoResLme}
\usage{
modelCathoResLme(d, group, age, res)
}
\arguments{
\item{d}{data.frame with data}

\item{group}{char with group names}

\item{age}{cell age}

\item{res}{cathode resistance value}
}
\value{
lmer model
}
\description{
Model cathode resistance per groups with  0 fixed intercept and varying (random)
intercept and slope using \code{lme4::lmer}
}
