% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCathRes.R
\name{modelCathoResLmeStan}
\alias{modelCathoResLmeStan}
\title{modelCathoResLmeStan}
\usage{
modelCathoResLmeStan(d, group, age, res)
}
\arguments{
\item{d}{data.frame with data}

\item{group}{char with group names}

\item{age}{cell age}

\item{res}{cathode resistance value}
}
\value{
Stan lmer model
}
\description{
Model cathode resistance per groups with  0 fixed intercept and varying (random)
intercept and slope
}
