% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCathRes.R
\name{plotCathRes}
\alias{plotCathRes}
\title{plotCathRes}
\usage{
plotCathRes(d, pot, age, res, ncol = 3)
}
\arguments{
\item{d}{datatable containing the data}

\item{pot}{character pot variable}

\item{age}{numeric pot age in days}

\item{res}{numeric cathode resistance in microhm}

\item{ncol}{number of columns for facetting}
}
\value{
ggplot with a facet for each pot
}
\description{
plots cathode resistance per pot for several pots  versus age
}
\examples{
data("resCathData")
d <- resCathData
plotCathRes(d, cod_cuve, agebsq, rucv)


}
