% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLogFunction.R
\name{findLogFunction}
\alias{findLogFunction}
\title{findLogFunction}
\usage{
findLogFunction(d, age, res)
}
\arguments{
\item{d}{datatable containing the data}

\item{age}{numeric pot age in days}

\item{res}{numeric cathode resistance in microhm}
}
\value{
function
}
\description{
Calculates the log function interpolating cathode resistance data
resistance = a + b * log(age)
}
\examples{
data("resCathData")
d <- resCathData
findLogFunction(d, agebsq, rucv)


}
